fix(collavre): address PR #1278 review
- Extract duplicated RecurringTask args parsing into shared
  Collavre::Crons::RecurringTaskArguments (CronListService + notifier)
- OrphanedCronNotifier: set skip_default_user so the notice stays a
  system message (user nil) and never triggers AI orchestration even
  when the topic deleter is Current.user (Codex P1)
- TopicsController#destroy: wrap notifier in rescue so a notice failure
  never breaks the core deletion or the 'deleted' broadcast
- Add regression test for the system-message attribution
